fix state writing and delete unused
